+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 This file contains changelog of [Dell Fan Monitor](https://github.com/ru-ace/dell-fan-mon/)
 ## [Unreleased]
 ### Added
-- **discrete_gpu_mode** for laptops with discrete gpu . 
+- **discrete_gpu_mode** for laptops with discrete gpu
+- **get_gpu_temp_cmd** for laptops with discrete gpu in case when gpu_temp_sensor_id autodetect failed.
 ### Changed
-- Set direct SMM BIOS calls (**mode 1**) by default (main cause: default **discrete_gpu_mode 1** need **mode 1**)
+- Set direct SMM BIOS calls (**mode 1**) by default (main cause: default **discrete_gpu_mode 1** with autodetect needs **mode 1**)
 - Doesn't checks and corrects fan state if **bios_disable_method** used. Considered that **bios_disable_method** works and fan state setted only by dell-fan-mon. 
 <!--## [Unreleased]